[1/n][object runtime type tags][cleanup-only] Pull `SuiDataStore` into separate module and cleanup import formatting
