+++ sui-execution/latest/sui-adapter/src/programmable_transactions/execution.rs
@@ -5,18 +5,21 @@ pub use checked::*;
 
 #[sui_macros::with_checked_arithmetic]
 mod checked {
-    use crate::execution_mode::ExecutionMode;
-    use crate::execution_value::{
-        CommandKind, ExecutionState, ObjectContents, ObjectValue, RawValueType, Value,
-        ensure_serialized_size,
+    use crate::{
+        adapter::substitute_package_id,
+        execution_mode::ExecutionMode,
+        execution_value::{
+            CommandKind, ExecutionState, ObjectContents, ObjectValue, RawValueType, Value,
+            ensure_serialized_size,
+        },
+        gas_charger::GasCharger,
+        programmable_transactions::{context::*, data_store::SuiDataStore},
     };
-    use crate::gas_charger::GasCharger;
-    use move_binary_format::file_format::AbilitySet;
     use move_binary_format::{
         CompiledModule,
         compatibility::{Compatibility, InclusionCheck},
         errors::{Location, PartialVMResult, VMResult},
-        file_format::{CodeOffset, FunctionDefinitionIndex, LocalIndex, Visibility},
+        file_format::{AbilitySet, CodeOffset, FunctionDefinitionIndex, LocalIndex, Visibility},
         file_format_common::VERSION_6,
         normalized,
     };
@@ -33,9 +36,8 @@ mod checked {
     };
     use move_vm_types::loaded_data::runtime_types::{CachedDatatype, Type};
     use serde::{Deserialize, de::DeserializeSeed};
-    use std::cell::OnceCell;
     use std::{
-        cell::RefCell,
+        cell::{OnceCell, RefCell},
         collections::{BTreeMap, BTreeSet},
         fmt,
         rc::Rc,
@@ -44,11 +46,6 @@ mod checked {
     };
     use sui_move_natives::object_runtime::ObjectRuntime;
     use sui_protocol_config::ProtocolConfig;
-    use sui_types::execution::{ExecutionTiming, ResultWithTimings};
-    use sui_types::execution_config_utils::to_binary_config;
-    use sui_types::execution_status::{CommandArgumentError, PackageUpgradeError};
-    use sui_types::storage::{PackageObject, get_package_objects};
-    use sui_types::type_input::TypeInput;
     use sui_types::{
         SUI_FRAMEWORK_ADDRESS,
         base_types::{
@@ -58,23 +55,25 @@ mod checked {
         },
         coin::Coin,
         error::{ExecutionError, ExecutionErrorKind, command_argument_error},
+        execution::{ExecutionTiming, ResultWithTimings},
+        execution_config_utils::to_binary_config,
+        execution_status::{CommandArgumentError, PackageUpgradeError},
         id::RESOLVED_SUI_ID,
         metrics::LimitsMetrics,
         move_package::{
             MovePackage, UpgradeCap, UpgradePolicy, UpgradeReceipt, UpgradeTicket,
             normalize_deserialized_modules,
         },
+        storage::{PackageObject, get_package_objects},
         transaction::{Command, ProgrammableMoveCall, ProgrammableTransaction},
         transfer::RESOLVED_RECEIVING_STRUCT,
+        type_input::TypeInput,
     };
     use sui_verifier::{
         INIT_FN_NAME,
         private_generics::{EVENT_MODULE, PRIVATE_TRANSFER_FUNCTIONS, TRANSFER_MODULE},
     };
     use tracing::instrument;
-
-    use crate::adapter::substitute_package_id;
-    use crate::programmable_transactions::context::*;
 
     pub fn execute<Mode: ExecutionMode>(
         protocol_config: &ProtocolConfig,
