+++ sui-execution/latest/sui-adapter/src/programmable_transactions/mod.rs
@@ -2,5 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod context;
+pub mod data_store;
 pub mod execution;
 pub mod linkage_view;
